Kernel functions

Here are the current base kernel functions that were used to construct the src/kernels kernel functions

GPLinearODEMaker.matern52_kernel_base — Method

matern52_kernel_base(λ, δ)

Matern 5/2 kernel. Twice mean square differentiable

Arguments

λ::Number: The kernel lengthscale

δ::Number: The difference between the inputs (e.g. t1 - t2)

GPLinearODEMaker.pp_kernel_base — Method

pp_kernel_base(λ, δ)

Piecewise polynomial kernel. Twice mean square differentiable. Equation 4.21 in Rasmussen and Williams.

Arguments

λ::Number: The kernel lengthscale, and cutoff variable

δ::Number: The difference between the inputs (e.g. t1 - t2)

GPLinearODEMaker.rm52_kernel_base — Method

rq_kernel_base(hyperparameters, δ)

Rational Matern 5/2 kernel. Equivalent to adding together Matern 5/2 kernels with the inverse of the lengthscale (τ = M52_λ^-1) are distributed as a Gamma distribution of p(τ|α,μ) where α (sometimes written as k) is the shape parameter and μ is the mean of the distribution.

Arguments

hyperparameters::Vector: The kernel shape parameter and mean (i.e. [α, μ])

δ::Number: The difference between the inputs (e.g. t1 - t2)

GPLinearODEMaker.rq_kernel_base — Method

rq_kernel_base(hyperparameters, δ)

Rational Quadratic kernel. Equivalent to adding together SE kernels with the inverse square of the lengthscales (τ = SE_λ^-2) are distributed as a Gamma distribution of p(τ|α,μ) where α (sometimes written as k) is the shape parameter and μ is the mean of the distribution. When α→∞, the RQ is identical to the SE with λ = μ^-1/2.

Arguments

hyperparameters::Vector: The kernel shape parameter and mean (i.e. [α, μ])

δ::Number: The difference between the inputs (e.g. t1 - t2)

GPLinearODEMaker.se_kernel_base — Method

se_kernel_base(λ, δ)

Squared exonential GP kernel (~Gaussian). Infinitely mean square differentiable (a.k.a. very smooth).

Arguments

λ::Number: The kernel lengthscale

δ::Number: The difference between the inputs (e.g. t1 - t2)

All of the premade kernels that are included with GLOM (in src/kernels) were created with this example script